+++ schema.graphql
@@ -1,5 +1,5 @@
 # source: https://graphql.kiwi.com/
-# timestamp: Wed Apr 11 2018 15:18:39 GMT+0200 (CEST)
+# timestamp: Wed Apr 11 2018 17:25:33 GMT+0200 (CEST)
 
 schema {
   query: RootQuery
