+++ src/relay/environment.js
@@ -4,25 +4,26 @@ import { Environment, Network, RecordSource, Store } from 'relay-runtime';
 
 require('isomorphic-fetch');
 
-// used when smart FAQ installed as dependency
-const uri = 'https://graphql.kiwi.com';
+const buildFetchQuery = (token: string = '') => {
+  return function fetchQuery(operation, variables) {
+    return fetch(process.env.GRAPHQL_URI, {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+        Authorization: token,
+      },
+      body: JSON.stringify({
+        query: operation.text, // GraphQL text from input
+        variables,
+      }),
+    }).then(response => response.json());
+  };
+};
+const buildEnvironment = (token: string = '') => {
+  return new Environment({
+    network: Network.create(buildFetchQuery(token)),
+    store: new Store(new RecordSource()),
+  });
+};
 
-function fetchQuery(operation, variables) {
-  return fetch(process.env.GRAPHQL_URI || uri, {
-    method: 'POST',
-    headers: {
-      'content-type': 'application/json',
-      Authorization:
-        'WyJnUDVIdmdlVURmZlU4MVVjVWtBYW4xIiwiQzR0RzR1ZHk3dHdkbDBSeEpQbWFadUl1MnhSUS92NnhCTFJzSS51ZGhoMXAySTduRk5KQVciLDIxMTU4MDg2NjZd.2q4_W34IS8YCX4PshJuKQkZt8VQ',
-    },
-    body: JSON.stringify({
-      query: operation.text, // GraphQL text from input
-      variables,
-    }),
-  }).then(response => response.json());
-}
-
-export default new Environment({
-  network: Network.create(fetchQuery),
-  store: new Store(new RecordSource()),
-});
+export default buildEnvironment;
